Factor out the log console status bar component to a separate module.
